HotelBackend: created function that test if a membership ID is valid
