+++ src/main/java/mainhotelapp/SysConstants.java
@@ -14,6 +14,13 @@ public class SysConstants {
     public enum RoomTypeConst {
         SUITE, REG, HANDI
     }
+
+
+
+
+
+
+
